+++ CHANGELOG.md
@@ -4,8 +4,11 @@
 
 ### Bug fixes
 
+- [Storybook] integrated for the demo.
 - Documentation mistakes fixed
 
 ## **[1.0.0]** - 16/01/2018
+
+[Storybook]: https://github.com/storybooks/storybook
 
 [1.0.0]: https://github.com/nitsnets/unicorn_components/releases/tag/v1.0.0 "v1.0.0"